feat: add editTodoData logic
